+++ sdk/cosmos/azure-cosmos/azure/cosmos/aio/_inference_service_async.py
@@ -40,7 +40,7 @@ from .. import exceptions
 from .._constants import _Constants as Constants
 from .._cosmos_http_logging_policy import CosmosHttpLoggingPolicy
 from .._cosmos_responses import CosmosDict
-from .._response_decoding import decode_response_body
+from .._response_decoding import decode_response_body_for_status
 from ..http_constants import HttpHeaders
 
 
@@ -236,7 +236,9 @@ class _InferenceService:
 
             data = response.body()
             if data:
-                data = decode_response_body(data, "inference_request")
+                data = decode_response_body_for_status(
+                    data, response.status_code, "inference_request"
+                )
 
             if response.status_code >= 400:
                 raise exceptions.CosmosHttpResponseError(message=data, response=response)